Simplify and optimize code
The for loop was very dumb since all it did was check if the current skin was within range of 0 to `amountOfSkins` - 1. Since `ChangeSkin.cs` is a component on player we don't actually need to have the player as variable we can just use GetComponent directly. RequireComponent was also added since guarantees that we have an animator. Skin 1 was broken before this.